+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - node.js v14+ required.
 - Removed `.run()` method. Use async/await.
 - `.exit()`: removed callback support. Use async/await.
+- Removed callback support in `.fn()` and all other functions, pass-able to
+  params. Use async/await.